Post Validation

+++ BlogApp/Services/PostService.cs
@@ -16,17 +16,32 @@ namespace BlogApp.Services
 
         public async Task AddPostAsync(PostDTO postDTO)
         {
-            var Post = new Post
-            {
-                Title = postDTO.Title,
-                Content = postDTO.Content,
-                CreatedAt = DateTime.Now,
-            };
-            await _postRepository.AddPostAsync(Post);
+            if (string.IsNullOrWhiteSpace(postDTO.Title) || postDTO.Title.Length > 100)
+            
+                throw new ArgumentException("Title cant be empty and must be shorter than 100 letters");
+
+            if (string.IsNullOrWhiteSpace(postDTO.Content) || postDTO.Content.Length < 10)
+
+                throw new ArgumentException("Content cant be empty and must be longer than 10 letters ");
+
+            
+                var Post = new Post
+                {
+                    Title = postDTO.Title,
+                    Content = postDTO.Content,
+                    CreatedAt = DateTime.Now,
+                };
+                await _postRepository.AddPostAsync(Post);
+            
         }
 
         public async Task DeletePostAsync(int id)
         {
+            var postExist = await _postRepository.GetPostAsync(id);
+            if (postExist == null)
+            {
+                throw new ArgumentException("Post doesnt exist");
+            }
             var post = await _postRepository.GetPostAsync(id);
             await _postRepository.DeletePostAsync(post);
         }
